relocate EventLikeButton
